requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-jose[crytography]
passlib[bcrypt]
bcrypt
motor
beanie
minio